Replace drop rates with absolute packet count check

+++ pkg/internal/checkup/executor/trex_console_types.go
@@ -5,7 +5,7 @@ type globalStats struct {
 	Jsonrpc string `json:"jsonrpc"`
 	Result  struct {
 		MActiveFlows            float64 `json:"m_active_flows"`
-		MActiveSockets          int     `json:"m_active_sockets"`
+		MActiveSockets          int64   `json:"m_active_sockets"`
 		MBwPerCore              float64 `json:"m_bw_per_core"`
 		MCPUUtil                float64 `json:"m_cpu_util"`
 		MCPUUtilRaw             float64 `json:"m_cpu_util_raw"`
@@ -17,22 +17,22 @@ type globalStats struct {
 		MRxDropBps              float64 `json:"m_rx_drop_bps"`
 		MRxPps                  float64 `json:"m_rx_pps"`
 		MSocketUtil             float64 `json:"m_socket_util"`
-		MTotalAllocError        int     `json:"m_total_alloc_error"`
-		MTotalClients           int     `json:"m_total_clients"`
-		MTotalNatActive         int     `json:"m_total_nat_active "`
-		MTotalNatLearnError     int     `json:"m_total_nat_learn_error"`
-		MTotalNatNoFid          int     `json:"m_total_nat_no_fid "`
-		MTotalNatOpen           int     `json:"m_total_nat_open   "`
-		MTotalNatSynWait        int     `json:"m_total_nat_syn_wait"`
-		MTotalNatTimeOut        int     `json:"m_total_nat_time_out"`
-		MTotalNatTimeOutWaitAck int     `json:"m_total_nat_time_out_wait_ack"`
-		MTotalQueueDrop         int     `json:"m_total_queue_drop"`
-		MTotalQueueFull         int     `json:"m_total_queue_full"`
-		MTotalRxBytes           int     `json:"m_total_rx_bytes"`
-		MTotalRxPkts            int     `json:"m_total_rx_pkts"`
-		MTotalServers           int     `json:"m_total_servers"`
-		MTotalTxBytes           int     `json:"m_total_tx_bytes"`
-		MTotalTxPkts            int     `json:"m_total_tx_pkts"`
+		MTotalAllocError        int64   `json:"m_total_alloc_error"`
+		MTotalClients           int64   `json:"m_total_clients"`
+		MTotalNatActive         int64   `json:"m_total_nat_active "`
+		MTotalNatLearnError     int64   `json:"m_total_nat_learn_error"`
+		MTotalNatNoFid          int64   `json:"m_total_nat_no_fid "`
+		MTotalNatOpen           int64   `json:"m_total_nat_open   "`
+		MTotalNatSynWait        int64   `json:"m_total_nat_syn_wait"`
+		MTotalNatTimeOut        int64   `json:"m_total_nat_time_out"`
+		MTotalNatTimeOutWaitAck int64   `json:"m_total_nat_time_out_wait_ack"`
+		MTotalQueueDrop         int64   `json:"m_total_queue_drop"`
+		MTotalQueueFull         int64   `json:"m_total_queue_full"`
+		MTotalRxBytes           int64   `json:"m_total_rx_bytes"`
+		MTotalRxPkts            int64   `json:"m_total_rx_pkts"`
+		MTotalServers           int64   `json:"m_total_servers"`
+		MTotalTxBytes           int64   `json:"m_total_tx_bytes"`
+		MTotalTxPkts            int64   `json:"m_total_tx_pkts"`
 		MTxBps                  float64 `json:"m_tx_bps"`
 		MTxCps                  float64 `json:"m_tx_cps"`
 		MTxExpectedBps          float64 `json:"m_tx_expected_bps"`
@@ -46,16 +46,16 @@ type portStats struct {
 	ID      string `json:"id"`
 	Jsonrpc string `json:"jsonrpc"`
 	Result  struct {
-		Ibytes      int     `json:"ibytes"`
-		Ierrors     int     `json:"ierrors"`
-		Ipackets    int     `json:"ipackets"`
+		Ibytes      int64   `json:"ibytes"`
+		Ierrors     int64   `json:"ierrors"`
+		Ipackets    int64   `json:"ipackets"`
 		MCPUUtil    float64 `json:"m_cpu_util"`
 		MTotalRxBps float64 `json:"m_total_rx_bps"`
 		MTotalRxPps float64 `json:"m_total_rx_pps"`
 		MTotalTxBps float64 `json:"m_total_tx_bps"`
 		MTotalTxPps float64 `json:"m_total_tx_pps"`
-		Obytes      int     `json:"obytes"`
-		Oerrors     int     `json:"oerrors"`
-		Opackets    int     `json:"opackets"`
+		Obytes      int64   `json:"obytes"`
+		Oerrors     int64   `json:"oerrors"`
+		Opackets    int64   `json:"opackets"`
 	} `json:"result"`
 }
